requirements.txt
flask>=2.3.0
requests>=2.28.0
pdfplumber>=0.9.0
pypdf>=3.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
